#19 - view mapping bug fix
requestMapping으로 controller별 uri 매핑
