+++ spring-mermer-shop-mvc/src/main/java/com/mermer/cm/util/ResourceServerConfig.java
@@ -37,7 +37,8 @@ public class ResourceServerConfig extends ResourceServerConfigurerAdapter{
 				.antMatchers(HttpMethod.POST,"/api/account").permitAll()
 				.antMatchers(HttpMethod.GET,"/api/notice/**").permitAll()
 				.antMatchers(HttpMethod.GET,"/api/bulletin/**").permitAll()
-				.anyRequest().authenticated()
+				.antMatchers("/api/**").authenticated()
+				.anyRequest().permitAll()
 				.and()
 			.formLogin()
 				.loginPage("/login")
